+++ lib/tests/test_revset_optimized.rs
@@ -104,6 +104,9 @@ fn arb_expression(
                 expr.clone().prop_map(|x| x.children()),
                 expr.clone().prop_map(|x| x.descendants()),
                 (expr.clone(), 0..5_u64).prop_map(|(x, d)| x.descendants_range(0..d)),
+                // First ancestors
+                expr.clone().prop_map(|x| x.first_ancestors()),
+                (expr.clone(), 0..5_u64).prop_map(|(x, d)| x.first_ancestors_range(0..d)),
                 // Range
                 (expr.clone(), expr.clone()).prop_map(|(x, y)| x.range(&y)),
                 // DagRange